Refactor result storage
With proper composite pattern this simplifies many places marked with NOT SO NICE.
The score of a single try is now really calculated in SingleResult.

+++ srcTools/magma/tools/benchmark/view/bench/goaliechallenge/GoalieBenchmarkTableModelExtension.java
@@ -69,7 +69,7 @@ class GoalieBenchmarkTableModelExtension extends DefaultTableModel
 	{
 		super(data, columnNames);
 
-		columnTypes = new Class[] {String.class, Object.class, Float.class, Integer.class, String.class};
+		columnTypes = new Class[] {String.class, Object.class, Double.class, Integer.class, String.class};
 
 		editableColumns = new boolean[] {true, false, false, false, true};
 	}
