+++ srcTools/magma/tools/benchmark/view/bench/keepawaychallenge/KeepAwayBenchmarkTableModelExtension.java
@@ -44,7 +44,7 @@ class KeepAwayBenchmarkTableModelExtension extends DefaultTableModel
 	{
 		super(data, columnNames);
 
-		columnTypes = new Class[] {String.class, Object.class, Float.class, String.class, Float.class};
+		columnTypes = new Class[] {String.class, Object.class, Double.class, String.class, Float.class};
 
 		editableColumns = new boolean[] {true, false, false, true, true};
 	}
