+++ srcTools/magma/tools/benchmark/view/bench/passingchallenge/PassingBenchmarkTableModelExtension.java
@@ -51,7 +51,7 @@ public class PassingBenchmarkTableModelExtension extends DefaultTableModel
 	{
 		super(data, columnNames);
 
-		columnTypes = new Class[] {String.class, Object.class, Float.class, Float.class, Float.class, Float.class,
+		columnTypes = new Class[] {String.class, Object.class, Double.class, Double.class, Double.class, Double.class,
 				String.class, Float.class};
 
 		editableColumns = new boolean[] {true, false, false, false, false, false, true, true};
