+++ srcTools/magma/tools/benchmark/view/bench/rollingballchallenge/RollingBallBenchmarkTableModelExtension.java
@@ -79,7 +79,7 @@ class RollingBallBenchmarkTableModelExtension extends DefaultTableModel
 	{
 		super(data, columnNames);
 
-		columnTypes = new Class[] {String.class, Object.class, Float.class, Integer.class, Float.class, Float.class,
+		columnTypes = new Class[] {String.class, Object.class, Double.class, Integer.class, Float.class, Float.class,
 				String.class, Float.class};
 
 		editableColumns = new boolean[] {true, false, false, false, false, false, true, true};
