+++ srcTools/magma/tools/benchmark/view/bench/runchallenge/RunBenchmarkTableModelExtension.java
@@ -88,8 +88,8 @@ class RunBenchmarkTableModelExtension extends DefaultTableModel
 	{
 		super(data, columnNames);
 
-		columnTypes = new Class[] {String.class, Object.class, Float.class, Integer.class, Integer.class, Float.class,
-				Float.class, Float.class, Float.class, String.class, Float.class};
+		columnTypes = new Class[] {String.class, Object.class, Double.class, Integer.class, Integer.class, Double.class,
+				Double.class, Double.class, Double.class, String.class, Float.class};
 
 		editableColumns = new boolean[] {true, false, false, false, false, false, false, false, false, true, true};
 	}
